trimed_backend/trimed_backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==24.2
Pillow==10.4.0
whitenoise==6.6.0
django-extensions==3.2.3
gunicorn==21.2.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
channels==4.1.0
daphne==4.1.0
pyotp==2.9.0
reportlab==4.1.0
django-sendgrid-v5==0.8.1
phonenumbers==8.13.22
pytest==7.4.4
dj-database-url==2.0.0
pytest-django==4.8.0